backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
duckduckgo-search==6.3.7
trafilatura==1.10.0
sentence-transformers==3.1.1
pinecone-client==5.0.1
transformers==4.45.1
torch>=2.0
spacy==3.7.2
requests
